added pffft

+++ tests/benchmarks/BenchFFTpffft.cpp
@@ -1,5 +1,5 @@
 #define TKLB_MAXCHANNELS 16
-#include "../../types/audio/fft/TOouraFFT.h"
+#include "../../types/audio/fft/Tpffft.h"
 #include "BenchmarkCommon.h"
 
 int main() {
@@ -7,7 +7,7 @@ int main() {
 		const int fftSize = 512;
 		FFT con = { fftSize };
 		AudioBuffer<> input, output;
-		AudioBuffer<double> result; // ooura only does doubles
+		AudioBuffer<float> result;
 		input.resize(fftSize, 1);
 		output.resize(fftSize, 1);
 		result.resize(fftSize / 2, 2);
